checkin/pyproject.toml
[project]
name = "iliad-checkin"
version = "0.1.0"
description = "Check-in tool for ILIAD OGC Building Blocks: register a data source, match bblocks, map vocabularies, generate a tested transformer and staged bblock."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "python-multipart>=0.0.9",
    "pydantic>=2.6",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "jsonschema>=4.21",
    "rdflib>=7.0",
    "pandas>=2.1",
    "pyarrow>=15.0",
    "xarray>=2024.1",
    "netCDF4>=1.6",
    "rapidfuzz>=3.6",
    "OWSLib>=0.31",
    "click>=8.1",
    "rich>=13.7",
]

[project.scripts]
iliad-checkin = "checkin.cli:main"
iliad-checkin-server = "checkin.api.server:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["checkin*"]
